Factor unnecessary util module.
